added timer and players
a timer has been added to check the correctness - after setting the color, 2 seconds pass and then it is checked whether the color is in the right place. If not - the turn passes to the next player and the box becomes blank. If player guessed correctly, the color stays in the middle row, the guessed color disappears from the color palette and the player has another try